requirements.txt
Django==3.0.6
psycopg2-binary==2.8.4
django-environ==0.4.5
django-allauth==0.41.0
pillow==6.0.0
django-markdownx==3.0.1
